+++ packages/antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.12.30
+
+### Patch Changes
+
+- update
+
 ## 2.12.29
 
 ### Patch Changes